Moved symbolics from solver to models.
