+++ pycollocation/solvers.py
@@ -1,7 +1,4 @@
-import collections
-
 import numpy as np
-import sympy as sym
 
 from . import models
 
@@ -9,15 +6,7 @@ from . import models
 class Solver(object):
     """Base class for all Solvers."""
 
-    __lower_boundary_condition = None
-
-    __upper_boundary_condition = None
-
-    _cached_rhs_functions = {}  # not sure if this is good practice!
-
-    _modules = [{'ImmutableMatrix': np.array}, 'numpy']
-
-    def __init__(self, model, params):
+    def __init__(self, model):
         """
         Create an instance of the Solver class.
 
@@ -25,49 +14,9 @@ class Solver(object):
         ----------
         model : models.Model
             An instance of models.Model to solve.
-        params : dict
-            A dictionary of model parameters.
 
         """
         self.model = model
-        self.params = params
-
-    @property
-    def _lower_boundary_condition(self):
-        """Cache lambdified lower boundary condition for numerical evaluation."""
-        condition = self.model.boundary_conditions['lower']
-        if condition is not None:
-            if self.__lower_boundary_condition is None:
-                self.__lower_boundary_condition = self._lambdify_factory(condition)
-            return self.__lower_boundary_condition
-        else:
-            return None
-
-    @property
-    def _symbolic_args(self):
-        """List of symbolic arguments used to lambdify expressions."""
-        return self._symbolic_vars + self._symbolic_params
-
-    @property
-    def _symbolic_params(self):
-        """List of symbolic model parameters."""
-        return sym.var(list(self.params.keys()))
-
-    @property
-    def _symbolic_vars(self):
-        """List of symbolic model variables."""
-        return [self.model.independent_var] + self.model.dependent_vars
-
-    @property
-    def _upper_boundary_condition(self):
-        """Cache lambdified upper boundary condition for numerical evaluation."""
-        condition = self.model.boundary_conditions['upper']
-        if condition is not None:
-            if self.__upper_boundary_condition is None:
-                self.__upper_boundary_condition = self._lambdify_factory(condition)
-            return self.__upper_boundary_condition
-        else:
-            return None
 
     @property
     def coefficients(self):
@@ -118,24 +67,6 @@ class Solver(object):
     def model(self, model):
         """Set a new model to solve."""
         self._model = self._validate_model(model)
-
-    @property
-    def params(self):
-        """
-        Dictionary of model parameters.
-
-        :getter: Return the current parameter dictionary.
-        :setter: Set a new parameter dictionary.
-        :type: dict
-
-        """
-        return self._params
-
-    @params.setter
-    def params(self, value):
-        """Set a new parameter dictionary."""
-        valid_params = self._validate_params(value)
-        self._params = self._order_params(valid_params)
 
     @property
     def result(self):
@@ -232,10 +163,10 @@ class Solver(object):
 
     def _evaluate_lower_boundary_residual(self, basis_funcs, lower_bound):
         """Return the lower boundary condition evaluated on the domain."""
-        if self._lower_boundary_condition is not None:
+        if self.model._lower_boundary_condition is not None:
             args = (self._evaluate_basis_funcs(basis_funcs, lower_bound) +
-                    list(self.params.values()))
-            return self._lower_boundary_condition(lower_bound, *args)
+                    list(self.model.params.values()))
+            return self.model._lower_boundary_condition(lower_bound, *args)
 
     def _evaluate_residual_funcs(self, residual_funcs, nodes):
         """Return a list of residual functions evaluated at collocation nodes."""
@@ -243,51 +174,27 @@ class Solver(object):
 
     def _evaluate_upper_boundary_residual(self, basis_funcs, upper_bound):
         """Return the upper boundary condition evaluated on the domain."""
-        if self._upper_boundary_condition is not None:
+        if self.model._upper_boundary_condition is not None:
             args = (self._evaluate_basis_funcs(basis_funcs, upper_bound) +
-                    list(self.params.values()))
-            return self._upper_boundary_condition(upper_bound, *args)
-
-    def _lambdify_factory(self, expr):
-        """Lambdify a symbolic expression."""
-        return sym.lambdify(self._symbolic_args, expr, self._modules)
+                    list(self.model.params.values()))
+            return self.model._upper_boundary_condition(upper_bound, *args)
 
     def _residual_function_factory(self, var, basis_derivs, basis_funcs):
         """Generate the residual function for a given variable."""
 
         def residual_function(t):
             """Residual function evaluated at array of points t."""
-            args = self._evaluate_basis_funcs(basis_funcs, t) + list(self.params.values())
-            return basis_derivs[var](t) - self._rhs_functions(var)(t, *args)
+            args = (self._evaluate_basis_funcs(basis_funcs, t) +
+                    list(self.model.params.values()))
+            return basis_derivs[var](t) - self.model._rhs_functions(var)(t, *args)
 
         return residual_function
-
-    def _rhs_functions(self, var):
-        """Cache lamdified rhs functions for numerical evaluation."""
-        if self._cached_rhs_functions.get(var) is None:
-            eqn = self.model.rhs[var]
-            self._cached_rhs_functions[var] = self._lambdify_factory(eqn)
-        return self._cached_rhs_functions[var]
-
-    @staticmethod
-    def _order_params(params):
-        """Cast a dictionary to an order dictionary."""
-        return collections.OrderedDict(sorted(params.items()))
 
     @staticmethod
     def _validate_model(model):
         """Validate the dictionary of parameters."""
-        if not isinstance(model, models.SymbolicModel):
-            mesg = "Attribute 'model' must have type models.SymbolicModel, not {}"
+        if not isinstance(model, models.BoundaryValueProblem):
+            mesg = "Attribute 'model' must have type models.BoundaryValueProblem, not {}"
             raise AttributeError(mesg.format(model.__class__))
         else:
             return model
-
-    @staticmethod
-    def _validate_params(value):
-        """Validate the dictionary of parameters."""
-        if not isinstance(value, dict):
-            mesg = "Attribute 'params' must have type dict, not {}"
-            raise AttributeError(mesg.format(value.__class__))
-        else:
-            return value
